Changed all three forms of the CParsedPhrase::ParsePhrase() function to be consistent on how the m_lstWords member is generated relative to the m_strCursorWord and m_nCursorWord values, as the text string and string list formats weren't appending the ending empty string, causing m_nCursorWord to possibly be outside the m_lstWords bounds. Changed m_nLevel/GetMatchLevel() and m_nCursorLevel/GetCursorMatchLevel() and phraseSize() and phraseSizeRaw() functions to just be signed like all of the other values to eliminate casts where we are using them.  Soon we'll be redefining how levels and phrasing works anyway to accommodate the "OR" operator, which will be implemented as subphrases within a phrase.
git-svn-id: file:///home/dewhisna/Documents/programs/Bible/svndump/purebiblesearch_150731/code_gitcvt/trunk@649 523e89ec-37fb-4b40-b97f-42a194c95123

+++ programs/KJVCanOpener/KJVSearchPhraseEdit.cpp
@@ -284,8 +284,8 @@ void CPhraseLineEdit::UpdateCompleter()
 	do {
 		cursor.selectWordUnderCursor();
 		if (/* (GetCursorWordPos() != nWord) && */
-			(static_cast<int>(GetMatchLevel()) <= nWord) &&
-			(static_cast<int>(GetCursorMatchLevel()) <= nWord) &&
+			(GetMatchLevel() <= nWord) &&
+			(GetCursorMatchLevel() <= nWord) &&
 			((nWord != GetCursorWordPos()) ||
 			 ((!GetCursorWord().isEmpty()) && (nWord == GetCursorWordPos()))
 			 )
